+++ molten_jwt/utils.py
@@ -1,6 +1,13 @@
-from molten import Header
+from molten import Header, Cookies
 
 from .exceptions import AuthenticationError
+
+
+def get_token_from_cookie(cookies: Cookies, cookie_name: str):
+    token = cookies.get(cookie_name)
+    if token is None:
+        raise AuthenticationError("Authorization cookie missing")
+    return token
 
 
 def get_token_from_header(authorization: Header, authorization_prefix: str):
@@ -13,6 +20,3 @@ def get_token_from_header(authorization: Header, authorization_prefix: str):
     if scheme.lower() != authorization_prefix:
         raise AuthenticationError("Authorization scheme not supported, try Bearer")
     return token
-
-
-# TODO create a get_token_from_cookie function